requirements.txt
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Mail==0.9.1
Flask-Limiter==3.5.0
WTForms==3.2.1
email-validator==2.1.1
python-dotenv==1.0.0
itsdangerous>=2.2.0
requests==2.31.0
SQLAlchemy==2.0.44
greenlet==3.2.4
typing_extensions==4.15.0
google-api-python-client==2.108.0

